Clean proj-dqry repo on branch ch/dev_clean (#3)
* updated gitignore file

* Modify gitignore file

* Merge config folder

* Modify ignore file - input folder

* Working trne workflow

* Reorganize git repo proj-dqry

Make sure the scripts are the most recent versions and organized them with explicit names and folders organization.

* Setup requirements for venv

Addtion of scikit-learn library to the requirements. It is required for the prediction-filter.py. Creation of requirement.in file. The requirements.txt libraries list is set to the one of the object-detector (2022-10)

* Update READ.md file

Add detailed description of the workflow procedure to performe the Training and Evaluation procedure and the Prediction procedure. Work still ongoing

* Update README.md

* Remove folder interface_proj-tpnl

* Remove folder interface_proj-dqry

* Modify .gitignore

Remove input and ignore folders from the list

* Update README.md

Move the file from src folder to main proj-dqry folder. Add the Copyright and License section

* Fix .gitignore file

Update the ignored files that will not be visible anymore on the repo

* Implement pull request sugestions

Implement changes suggested for pull request Clean proj-dqry repo on branch ch/dev_clean. Remove empty folder input. Add config folder. Rename src to scripts. Mv prost-processing and pre-processing one level up. Store images in image folders

* Correct the doc images path

Creation of an images folder in which all the images inclluded in the README.md files are stored. The respective path to the images in the README.md files have been updated.