Re-organisation of README.md files
Completed information to run projcet. 3 README.md files have been written. One added to the main project to give an overview and main infos. One added to the scripts folder to described each script functions. One added to the config folder to explain in detail how to run the workflow.